Removed header right sidebar area
After working with Compass for a while, making this a sidebar area
doesn’t really make a lot of sense. This was a concept brought over
from Genesis and it’s not one that works well in practice. I’m leaving
the template in for anyone who would like to use it in their projects,
but the widget area will no longer be enabled by default.