Switched Hybrid Core & Flagship Library to dev
There are some new changes coming in Hybrid Core that we need to be out
in front of for the 2.0.0 release, so I’m switching to the dev branch
until the next version of hybrid core is released. The next version of
the Flagship Library will be released soon so that change is temporary.